requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
idna==3.3
PyHackTheBox==0.5.6.post1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.11
